+++ src/qtasks/brokers/async_rabbitmq.py
@@ -345,5 +345,6 @@ class AsyncRabbitMQBroker(BaseBroker):
         for plugin in self.plugins.values():
             await plugin.trigger(name=name, *args, **kwargs)
     
-    async def flush_all(self):
+    async def flush_all(self) -> None:
+        """Удалить все данные."""
         await self.storage.flush_all()