+++ src/qtasks/brokers/base.py
@@ -254,5 +254,6 @@ class BaseBroker(ABC):
         for plugin in self.plugins.values():
             plugin.trigger(name=name, *args, **kwargs)
     
-    def flush_all(self):
+    def flush_all(self) -> None:
+        """Удалить все данные."""
         pass