+++ src/qtasks/configs/config_observer.py
@@ -2,12 +2,42 @@ from typing import Any, Callable, Dict, List, get_type_hints
 from qtasks.configs.config import QueueConfig
 
 class ConfigObserver:
+    """
+    `ConfigObserver` - класс конфига, который хранит в себе `QueueConfig`.
+
+    ## Пример
+
+    ```python
+    from qtasks import QueueTasks
+    from qtasks.configs import ConfigObserver, QueueConfig
+    
+    app = QueueTasks()
+    
+    config_classes = QueueConfig()
+    config = ConfigObserver(config_classes)
+    app.config_classes = config_classes
+    app.config = config
+    ```
+    """
     def __init__(self, config: QueueConfig):
+        """Экземпляр класса.
+
+        Args:
+            config (QueueConfig): Конфиг.
+        """
+        
         self._config = config
+        """QueueConfig"""
+
         self._callbacks: List[Callable[[str, Any], None]] = []
         self._dynamic_fields: Dict[str, Any] = {}
 
     def subscribe(self, callback: Callable[[str, Any], None]):
+        """Подписка на изменение конфига.
+
+        Args:
+            callback (Callable[[str, Any], None]): Класс конфига.
+        """
         self._callbacks.append(callback)
 
     def _notify(self, key: str, value: Any):