+++ src/qtasks/contrib/redis/sync_queue_client.py
@@ -5,9 +5,28 @@ import redis
 from qtasks.logs import Logger
 
 class SyncRedisCommandQueue:
+    """
+    `SyncRedisCommandQueue` - Асинхронный класс для работы с `Redis`.
+
+    ## Пример
+
+    ```python
+    from qtasks import QueueTasks
+    from qtasks.contrib.redis import SyncRedisCommandQueue
+    
+    redis_contrib = SyncRedisCommandQueue(redis)
+    redis_contrib.execute("hset", kwargs["name"], mapping=kwargs["mapping"])
+    ```
+    """
     def __init__(self,
             redis: redis.Redis,
             log: Logger = None):
+        """Экземпляр класса.
+
+        Args:
+            redis (redis.asyncio.Redis): класс `Redis`.
+            log (Logger, optional): класс `qtasks.logs.Logger`. По умолчанию: `qtasks._state.log_main`.
+        """
         self.log = self._get_log(log)
         self.redis = redis
         self.queue = Queue()
@@ -27,7 +46,14 @@ class SyncRedisCommandQueue:
         with self.lock:
             self.worker_thread = None  # Отмечаем, что воркер завершился
 
-    def execute(self, cmd, *args, **kwargs):
+    def execute(self, cmd: str, *args, **kwargs):
+        """Запрос в `Redis`.
+
+        Args:
+            cmd (str): Команда.
+            args(tuple, optional): Параметры к команде через *args.
+            kwargs(dict, optional): Параметры к команде через *args.
+        """
         self.queue.put((cmd, args, kwargs))
         with self.lock:
             if self.worker_thread is None or not self.worker_thread.is_alive():