+++ src/qtasks/executors/sync_task_executor.py
@@ -14,6 +14,19 @@ if TYPE_CHECKING:
 
 
 class SyncTaskExecutor(BaseTaskExecutor):
+    """
+    `SyncTaskExecutor` - Синхронный класс исполнителя задач. Используется по умолчанию в `SyncThreadWorker`.
+
+    ## Пример
+
+    ```python
+    from qtasks.executors import SyncTaskExecutor
+    
+    task_func = TaskExecSchema(...)
+    task_broker = TaskPrioritySchema(...)
+    executor = SyncTaskExecutor(task_func, task_broker)
+    result = executor.execute()
+    """
     def __init__(self,
             task_func: Annotated[
                 TaskExecSchema,
@@ -52,6 +65,14 @@ class SyncTaskExecutor(BaseTaskExecutor):
                 )
             ] = None
         ):
+        """Инициализация класса. Происходит внутри `Worker` перед обработкой задачи.
+
+        Args:
+            task_func (TaskExecSchema): Схема `TaskExecSchema`.
+            task_broker (TaskPrioritySchema): Схема `TaskPrioritySchema`.
+            middlewares (List[Type[TaskMiddleware]], optional): _description_. По умолчанию `None`.
+            log (Logger, optional): класс `qtasks.logs.Logger`. По умолчанию: `qtasks._state.log_main`.
+        """
         super().__init__(task_func=task_func, task_broker=task_broker, middlewares=middlewares, log=log)
 
         self._args = []
@@ -60,6 +81,7 @@ class SyncTaskExecutor(BaseTaskExecutor):
 
 
     def before_execute(self):
+        """Вызов перед запуском задач."""
         if self.task_func.echo:
             echo = SyncTask(task_name=self.task_broker.name, priority=self.task_broker.priority,
                 echo=self.task_func.echo,
@@ -72,20 +94,27 @@ class SyncTaskExecutor(BaseTaskExecutor):
         self._kwargs = self.task_broker.kwargs
 
     def after_execute(self):
+        """Вызов после запуска задач."""
         pass
 
     def execute_middlewares(self):
+        """Вызов мидлварей."""
         for m in self.middlewares:
             m = m(self)
             self.log.debug(f"Middleware {m.name} для {self.task_func.name} был вызван.")
 
     def run_task(self) -> Any:
-        if self._args and self._kwargs:
-            result = self.task_func.func(*self._args, **self._kwargs)
-        elif self._args:
-            result = self.task_func.func(*self._args)
-        elif self._kwargs:
-            result = self.task_func.func(**self._kwargs)
+        """Вызов задачи.
+
+        Returns:
+            Any: Результат задачи.
+        """
+        if self.task_broker.args and self.task_broker.kwargs:
+            result = self.task_func(*self.task_broker.args, **self.task_broker.kwargs)
+        elif self.task_broker.args:
+            result = self.task_func.func(*self.task_broker.args)
+        elif self.task_broker.kwargs:
+            result = self.task_func.func(**self.task_broker.kwargs)
         else:
             result = self.task_func.func()
         return result
@@ -93,6 +122,14 @@ class SyncTaskExecutor(BaseTaskExecutor):
     def execute(self,
             decode: bool = True
         ) -> Any|str:
+        """Вызов задачи.
+
+        Args:
+            decode (bool, optional): Декодирование результата задачи. По умолчанию: True.
+
+        Returns:
+            Any|str: Результат задачи.
+        """
         self.log.debug(f"Вызван execute для {self.task_func.name}")
         self.before_execute()
         self.execute_middlewares()
@@ -103,4 +140,9 @@ class SyncTaskExecutor(BaseTaskExecutor):
         return self._result
 
     def decode(self) -> str:
+        """Декодирование задачи.
+
+        Returns:
+            str: Результат задачи.
+        """
         return json.dumps(self._result, ensure_ascii=False)