+++ src/qtasks/storages/async_redis.py
@@ -10,7 +10,7 @@ import redis.asyncio as aioredis
 from typing import TYPE_CHECKING
 
 from qtasks.configs.config_observer import ConfigObserver
-from qtasks.contrlib.redis.async_queue_client import AsyncRedisCommandQueue
+from qtasks.contrib.redis.async_queue_client import AsyncRedisCommandQueue
 from qtasks.enums.task_status import TaskStatusEnum
 from qtasks.logs import Logger
 
@@ -120,7 +120,7 @@ class AsyncRedisStorage(BaseStorage):
         self._queue_process = queue_process
         self.queue_process = f"{self.name}:{queue_process}"
         self.client = redis_connect or aioredis.from_url(self.url, decode_responses=True, encoding=u'utf-8')
-        self.redis_contrlib = AsyncRedisCommandQueue(redis=self.client, log=self.log)
+        self.redis_contrib = AsyncRedisCommandQueue(redis=self.client, log=self.log)
         
         self.global_config = global_config or AsyncRedisGlobalConfig(name=self.name, redis_connect=self.client, log=self.log, config=self.config)
         
@@ -223,7 +223,7 @@ class AsyncRedisStorage(BaseStorage):
         Args:
             kwargs (dict, optional): данные задачи типа kwargs.
         """
-        return await self.redis_contrlib.execute("hset", kwargs["name"], mapping=kwargs["mapping"])
+        return await self.redis_contrib.execute("hset", kwargs["name"], mapping=kwargs["mapping"])
         
     async def remove_finished_task(self,
             task_broker: Annotated[
@@ -254,8 +254,8 @@ class AsyncRedisStorage(BaseStorage):
             model = TaskStatusErrorSchema(task_name=task_broker.name, priority=task_broker.priority, traceback=trace, created_at=task_broker.created_at, updated_at=time.time())
             self.log.warning(f"Задача {task_broker.uuid} завершена с ошибкой:\n{trace}")
         
-        await self.redis_contrlib.execute("hset", f"{self.name}:{task_broker.uuid}", mapping=model.__dict__)
-        await self.redis_contrlib.execute("zrem", self.queue_process, f"{task_broker.name}:{task_broker.uuid}:{task_broker.priority}")
+        await self.redis_contrib.execute("hset", f"{self.name}:{task_broker.uuid}", mapping=model.__dict__)
+        await self.redis_contrib.execute("zrem", self.queue_process, f"{task_broker.name}:{task_broker.uuid}:{task_broker.priority}")
         return
     
     async def start(self):
@@ -318,7 +318,8 @@ class AsyncRedisStorage(BaseStorage):
             await self.client.delete(*tasks_hash)
         return
 
-    async def flush_all(self):
+    async def flush_all(self) -> None:
+        """Удалить все данные."""
         loop = asyncio.get_running_loop()
         asyncio_atexit.register(self.stop, loop=loop)